+++ mongo_db_restore.py
@@ -26,7 +26,8 @@ exit 2
 
     Usage:
         mongo_db_restore.py -c file -d path
-            {-S db_name -o path [-z] [-i] [-r] [-k] [-y] [-e]}
+            {-S db_name -o path [-z] [-i] [-r] [-k] [-y] [-e] |
+             -C db_name -t coll_name -o path [-z] [-i] [-k] [-y] [-e]}
             [-p path] [-y flavor_id]
             [-v | -h]
 
@@ -34,11 +35,20 @@ exit 2
         -c file => Server configuration file.
         -d dir path => Directory path to config file (-c).
 
-        -S db_name => Database name to be restored to.
-            -o dir path => Directory path to database dump directory.
+        -S db_name => Restore database - pass database name.
+            -o dir path/db_name => Directory path to database dump directory.
             -z => Uncompress dump files.
             -i => Turn off TLS checking.
             -r => Restore database users and roles.
+            -k => Drop and recreate collection before restore.
+            -u => Run a dryrun of the restore.
+            -e => Run in verbose mode.
+
+        -C coll_name => Restore collection - pass collection name.
+            -b db_name => Name of database.
+            -o dir path/db_name => Directory path to database dump directory.
+            -z => Uncompress dump files.
+            -i => Turn off TLS checking.
             -k => Drop and recreate collection before restore.
             -u => Run a dryrun of the restore.
             -e => Run in verbose mode.
@@ -169,6 +179,54 @@ def help_message():
     print(__doc__)
 
 
+def restore(server, args, **kwargs):
+
+    """Function:  restore
+
+    Description:  Run restore command in subprocess call.
+
+    Arguments:
+        (input) server -> Mongo server instance
+        (input) args -> ArgParser class instance
+        (input) **kwargs:
+            opt_arg -> Dictionary of additional options to add
+            req_arg -> List of options to add to cmd line
+
+    """
+
+    load_cmd = mongo_libs.create_cmd(
+        server, args, "mongorestore", "-p", no_pass=True, **kwargs)
+    proc2 = subprocess.Popen(                           # pylint:disable=R1732
+        ["echo", server.japd], stdout=subprocess.PIPE)
+
+    proc1 = subprocess.Popen(                           # pylint:disable=R1732
+        load_cmd, stdin=proc2.stdout)
+    proc1.wait()
+
+
+def single_collection(server, args, **kwargs):
+
+    """Function:  single_collection
+
+    Description:  Restore single collection.
+
+    Arguments:
+        (input) server -> Mongo server instance
+        (input) args -> ArgParser class instance
+        (input) **kwargs:
+            opt_arg -> Dictionary of additional options to add
+            req_arg -> List of options to add to cmd line
+
+    """
+
+    req_arg = list(kwargs.get("req_arg", []))
+    opt_arg = dict(kwargs.get("opt_arg", {}))
+
+    # Do not allow restoring of users or roles during a collection restore
+    del opt_arg["-r"]
+
+    restore(server, args, req_arg=req_arg, opt_arg=opt_arg)
+
 def single_db(server, args, **kwargs):
 
     """Function:  single_db
@@ -176,41 +234,40 @@ def single_db(server, args, **kwargs):
     Description:  Restore single database.
 
     Arguments:
-        (input) server -> Database server instance
+        (input) server -> Mongo server instance
         (input) args -> ArgParser class instance
         (input) **kwargs:
             opt_arg -> Dictionary of additional options to add
             req_arg -> List of options to add to cmd line
-        (output) status -> True|False - If an error has occurred and associated
-            error message
 
     """
 
-    status = (False, None)
-    auth_db = "--authenticationDatabase="
+    restore(server, args, **kwargs)
 
+#    status = (False, None)
 #    req_arg = list(kwargs.get("req_arg", []))
 #    opt_arg = dict(kwargs.get("opt_arg", {}))
+#    auth_db = "--authenticationDatabase="
 
 #    if auth_db in req_arg:
 #        req_arg.remove(auth_db)
 #        req_arg.append(auth_db + server.auth_db)
 
-    load_cmd = mongo_libs.create_cmd(
-        server, args, "mongorestore", "-p", no_pass=True, **kwargs)
-    proc2 = subprocess.Popen(                           # pylint:disable=R1732
-        ["echo", server.japd], stdout=subprocess.PIPE)
+#    load_cmd = mongo_libs.create_cmd(
+#        server, args, "mongorestore", "-p", no_pass=True, **kwargs)
+#    proc2 = subprocess.Popen(                           # pylint:disable=R1732
+#        ["echo", server.japd], stdout=subprocess.PIPE)
 
 #    load_cmd = mongo_libs.create_cmd(
 #        server, args, "mongorestore", "-p", req_arg=req_arg,
 #        opt_arg=opt_arg)
 
-    proc1 = subprocess.Popen(                           # pylint:disable=R1732
-        load_cmd, stdin=proc2.stdout)
-    proc1.wait()
+#    proc1 = subprocess.Popen(                           # pylint:disable=R1732
+#        load_cmd, stdin=proc2.stdout)
+#    proc1.wait()
 #    proc1 = subprocess.Popen(load_cmd)                  # pylint:disable=R1732
 
-    return status
+#    return status
 
 
 def get_req_options(server, arg_req_dict):
@@ -265,12 +322,11 @@ def run_program(args, func_dict, **kwargs):
 
         # Intersect args_array and func_dict to find which functions to call
         for item in set(args.get_args_keys()) & set(func_dict.keys()):
-            err_flag, err_msg = func_dict[item](
-                server, args, req_arg=req_arg, opt_arg=opt_arg)
+            func_dict[item](server, args, req_arg=req_arg, opt_arg=opt_arg)
 #            err_flag, err_msg = func_dict[item](server, args, **kwargs)
 
-            if err_flag:
-                print(err_msg)
+#            if err_flag:
+#                print(err_msg)
 
         mongo_libs.disconnect([server])
 
@@ -290,8 +346,10 @@ def main():
         dir_perms_chk -> contains directories and their octal permissions
         func_dict -> dictionary list for the function calls or other options
         opt_arg_list -> contains optional arguments for the command line
+        opt_con_req_list -> contains the options that require other options
         opt_req_list -> contains the options that are required for the program
         opt_val_list -> contains options which require values
+        opt_xor_dict -> contains dict with key that is xor with it's values
 #        req_arg_list -> contains arguments to add to command line by default
 
     Arguments:
@@ -301,13 +359,15 @@ def main():
 
     arg_req_dict = {"auth_db": "--authenticationDatabase="}
     dir_perms_chk = {"-d": 5, "-o": 5, "-p": 5}
-    func_dict = {"-S": single_db}
+    func_dict = {"-S": single_db, "-C": single_collection}
     opt_arg_list = {
         "-S": "--db=", "-o": "--dir=", "-z": "--gzip", "-i": "--tlsInsecure",
         "-r": "--restoreDbUsersAndRoles", "-k": "--drop", "-u": "--dryRun",
-        "-e": "--verbose"}
+        "-e": "--verbose", "-b": "--db=", "-C": "--collection="}
+    opt_con_req_list = {"-S": ["-o"], "-C": ["-o", "-b"]}
     opt_req_list = ["-c", "-d", "-o"]
-    opt_val_list = ["-c", "-d", "-o", "-p", "-S", "-y"]
+    opt_val_list = ["-c", "-d", "-o", "-p", "-S", "-y", "-b", "-C"]
+    opt_xor_dict = {"-S": ["-C"], "-C": ["-S"]}
 #    req_arg_list = ["--authenticationDatabase="]
 
     # Process argument list from command line
@@ -316,7 +376,9 @@ def main():
     if args.arg_parse2()                                            \
        and not gen_libs.help_func(args, __version__, help_message)  \
        and args.arg_require(opt_req=opt_req_list)                   \
-       and args.arg_dir_chk(dir_perms_chk=dir_perms_chk):
+       and args.arg_dir_chk(dir_perms_chk=dir_perms_chk)            \
+       and args.arg_xor_dict(opt_xor_val=opt_xor_dict)              \
+       and args.arg_cond_req(opt_con_req=opt_con_req_list):
 
         try:
             prog_lock = gen_class.ProgramLock(